페이지처리, 검색처리X

+++ worksapce/spring/MyCloset/src/main/java/com/mycloset/notice/mapper/NoticeMapper.java
@@ -26,6 +26,6 @@ public interface NoticeMapper {
 	
 	//5. 삭제(delete)
 	
-	public int delete(int no) throws Exception;
+	public int delete(long no) throws Exception;
 
 }
